[Refactor]ユーティリティフォルダを作成&該当ファイルをまとめた
