+++ frontend/src/components/GroupTabs.tsx
@@ -1,6 +1,6 @@
 import Tabs from "@mui/material/Tabs";
 import Tab from "@mui/material/Tab";
-import { Events, getEventKey, getGroupName, GroupNames } from "../Type";
+import { Events, getEventKey, getGroupName, GroupNames } from "../utilities/Type";
 import { useEffect } from "react";
 
 type GroupTabProps = {
@@ -9,24 +9,21 @@ type GroupTabProps = {
   setSelectGroup: React.Dispatch<React.SetStateAction<number>>;
 };
 
-const GroupTabs = ({
-  selectEvent,
-  selectGroup,
-  setSelectGroup,
-}: GroupTabProps) => {
+const GroupTabs = ({ selectEvent, selectGroup, setSelectGroup }: GroupTabProps) => {
   return (
     <div className="group-tabs">
-      {GroupNames[selectEvent] && Object.keys(GroupNames[selectEvent]).map((groupKey) => (
-        <div
-          key={groupKey}
-          className={`group-tabs__item ${
-            selectGroup === parseInt(groupKey) ? "group-tabs__item--active" : ""
-          }`}
-          onClick={() => setSelectGroup(parseInt(groupKey))}
-        >
-          {getGroupName(selectEvent, parseInt(groupKey))}
-        </div>
-      ))}
+      {GroupNames[selectEvent] &&
+        Object.keys(GroupNames[selectEvent]).map((groupKey) => (
+          <div
+            key={groupKey}
+            className={`group-tabs__item ${
+              selectGroup === parseInt(groupKey) ? "group-tabs__item--active" : ""
+            }`}
+            onClick={() => setSelectGroup(parseInt(groupKey))}
+          >
+            {getGroupName(selectEvent, parseInt(groupKey))}
+          </div>
+        ))}
     </div>
   );
 };
