+++ frontend/src/utilities/ElementUtil.ts
